+++ src/index.js
@@ -1,28 +1,6 @@
-import readline from "readline";
 import CarRace from "./domain/CarRace.js";
 import { createCars } from "./utils/cars.js";
-
-function readLineAsync(query) {
-  return new Promise((resolve, reject) => {
-    if (arguments.length !== 1) {
-      reject(new Error("arguments must be 1"));
-    }
-
-    if (typeof query !== "string") {
-      reject(new Error("query must be string"));
-    }
-
-    const rl = readline.createInterface({
-      input: process.stdin,
-      output: process.stdout,
-    });
-
-    rl.question(`${query}\n`, (input) => {
-      rl.close();
-      resolve(input);
-    });
-  });
-}
+import { readLineAsync } from "./utils/io.js";
 
 async function play() {
   const input = await readLineAsync(
