+++ src/utils/io.js
@@ -1,0 +1,23 @@
+import readline from "readline";
+
+export function readLineAsync(query) {
+  return new Promise((resolve, reject) => {
+    if (arguments.length !== 1) {
+      reject(new Error("arguments must be 1"));
+    }
+
+    if (typeof query !== "string") {
+      reject(new Error("query must be string"));
+    }
+
+    const rl = readline.createInterface({
+      input: process.stdin,
+      output: process.stdout,
+    });
+
+    rl.question(`${query}\n`, (input) => {
+      rl.close();
+      resolve(input);
+    });
+  });
+}
